added resource dir to maven xml, fixed UserManagerImplementation for unitTests (handling exceptions)

+++ src/main/java/cz/muni/fi/pv168/UserManagerImpl.java
@@ -6,6 +6,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.IllegalFormatException;
 import java.util.List;
 
 /**
@@ -21,11 +22,16 @@ public class UserManagerImpl implements UserManager {
 
     @Override
     public void createUser(User user) {
+        if(user == null) throw new IllegalArgumentException("user is null");
+        if(user.getId() != null) throw new IllegalArgumentException("creating user with existing id");
+        if(user.getEmail() == null) throw new IllegalArgumentException("user email is null");
+        if(user.getEmail().trim().isEmpty()) throw new IllegalArgumentException("user email is empty");
+
         try(Connection conn = dataSource.getConnection()) {
             try(PreparedStatement ps = conn.prepareStatement("INSERT INTO USERS (FULLNAME, EMAIL) VALUES (?,?)",
                                                              PreparedStatement.RETURN_GENERATED_KEYS)) {
                 ps.setString(1,user.getFullName());
-                ps.setString(2,user.getEmail());
+                ps.setString(2,user.getEmail().toLowerCase());
                 ps.executeUpdate();
 
                 try(ResultSet keys = ps.getGeneratedKeys() ) {
@@ -36,18 +42,29 @@ public class UserManagerImpl implements UserManager {
                 }
             }
         } catch(SQLException ex) {
+            if(ex.getErrorCode() == 30000) {
+                throw new IllegalArgumentException("user email already exists");
+            }
             ex.printStackTrace();
         }
     }
 
     @Override
     public void updateUser(User user) {
+        if(user == null) throw new IllegalArgumentException("user is null");
+        if(user.getId() == null) throw new IllegalArgumentException("updating user with null id");
+        if(user.getEmail() == null) throw new IllegalArgumentException("user email is null");
+        if(user.getEmail().trim().isEmpty()) throw new IllegalArgumentException("user email is empty");
+
         try(Connection conn = dataSource.getConnection()) {
-            try(PreparedStatement ps = conn.prepareStatement("UPDATE USERS SET FULLANEM=?, EMAIL=? WHERE ID = ?")) {
+            try(PreparedStatement ps = conn.prepareStatement("UPDATE USERS SET FULLNAME=?, EMAIL=? WHERE ID = ?")) {
                 ps.setString(1, user.getFullName());
                 ps.setString(2, user.getEmail());
                 ps.setLong(3, user.getId());
                 int n = ps.executeUpdate();
+                if(n == 0) {
+                    throw new IllegalArgumentException("updating user with nonexisting id");
+                }
                 if(n != 1) {
                     //TODO throw exception
                 }
@@ -60,11 +77,16 @@ public class UserManagerImpl implements UserManager {
 
     @Override
     public void deleteUser(User user) {
+        if(user == null) throw new IllegalArgumentException("user is null");
+        if(user.getId() == null) throw new IllegalArgumentException("user id is null");
 
         try(Connection conn = dataSource.getConnection()) {
             try(PreparedStatement ps = conn.prepareStatement("DELETE FROM USERS WHERE ID = ?")) {
                 ps.setLong(1, user.getId());
                 int n = ps.executeUpdate();
+                if(n == 0) {
+                    throw new IllegalArgumentException("deleting user with nonexisting id");
+                }
                 if(n != 1) {
                     //TODO throw exeption
                 }
@@ -79,7 +101,7 @@ public class UserManagerImpl implements UserManager {
         try(Connection conn = dataSource.getConnection()) {
             try(PreparedStatement ps = conn.prepareStatement("SELECT * FROM USERS WHERE ID = ?")) {
                 ps.setLong(1, id);
-                getUserFromQuery(ps);
+                return getUserFromQuery(ps);
             }
         } catch (SQLException ex) {
             ex.printStackTrace();
@@ -90,9 +112,9 @@ public class UserManagerImpl implements UserManager {
     @Override
     public User getUserByEmail(String email) {
         try(Connection conn = dataSource.getConnection()) {
-            try(PreparedStatement ps = conn.prepareStatement("SELECT * FROM USERS WHERE EMAIL = ?")) {
+            try(PreparedStatement ps = conn.prepareStatement("SELECT * FROM USERS WHERE LOWER(EMAIL) = LOWER(?)")) {
                 ps.setString(1, email);
-                getUserFromQuery(ps);
+                return getUserFromQuery(ps);
 
             }
         } catch (SQLException ex) {
@@ -104,7 +126,7 @@ public class UserManagerImpl implements UserManager {
     @Override
     public List<User> getUserByName(String name) {
         try(Connection conn = dataSource.getConnection()) {
-            try(PreparedStatement ps = conn.prepareStatement("SELECT * FROM USERS WHERE FULLNAME = ?")) {
+            try(PreparedStatement ps = conn.prepareStatement("SELECT * FROM USERS WHERE LOWER(FULLNAME) = LOWER(?)")) {
                 ps.setString(1, name);
                 return getUserListFromQuery(ps);
             }
